Mouse: window must move to get tiled
Previous behaviour caused issue on clicking the bar of the window. It would auto-tile it based on the pointer location, which may combine tile the window if it was just below another tile.

I fixed that behaviour with additional checks, but even then the UX felt wrong. If you clicked on a window and immediately released, the tile layout preview flashed on screen.

The situation is confused further because the top bars of some windows in moder apps aren't "proper", so they can add functionality, e.g. tabs & menus. These can then alter whether they propagate the grab event, so the preview on click hold becomes inconsistent for a few apps that don't propagate the grab until the first mouse move.

Overall it's just a better UX to not count a window as moving until the pointer is actually moved, not when the initial grab happens.