fix disabling aspect ratio window size hint
At some point recently (<=GNOME 47), the window description has stopped being the X window ID & now also includes the title. It now has the format:
"{xid} ({title})"
This meant the aspect ratio window size hint wasn't being updated. It also caused an exception to be raised if the window title contained a single quote, which got seen as an unterminated string when trying to spawn the command.

The fix is to extract the X window ID from the description, allowing for it to have the title (or not), before passing it through to the command.